Clear ComputedStyle before detaching descendants.
More importantly, make sure ComputedStyle is cleared before cancelling
animations. There is a problem with CSSAnimations::Cancel() marking for
style recalc during DetachLayoutTree() and updating the style recalc
root when DetachLayoutTree() was called due to an element being removed
from the flat tree during slot assignment.

At that time, the element is outside the flat tree for traversal, but
may have a ComputedStyle which is not IsEnsuredOutsideFlatTree(). The
code to avoid having style recalc roots outside the flat tree relies on
ComputedStyle being null or IsEnsuredOutsideFlatTree().

An alternative approach to clearing the ComputedStyle early, would be to
pass an argument to Cancel() saying that the element is being removed
from the flat tree and make sure we don't try to mark for style recalc
in that case. Or, introduce a DetachLayoutTreeNotForReattachScope which
sets a flag to make sure we don't mark for style recalc or update the
recalc root.

Bug: 1093747
Change-Id: I7c133e6372a11bd0b87002f4eb54316588d05564