Reland "Don't reset computed 'float' for flex and grid items"
This is a reland of 906abd41e4f6c786a1927f6edd69d6926e20317b

Original change :
https://chromium-review.googlesource.com/c/chromium/src/+/2157889

Revert of original change :
https://chromium-review.googlesource.com/c/chromium/src/+/2256772

-----

The original change was reverted due to a DCHECK being triggered. See
the following CRBug for more details:
https://bugs.chromium.org/p/chromium/issues/detail?id=1097595

The original change no longer adjusted the ComputedStyle for floats
based on whether or not it was a flex/grid item. This adjustment was
left for the LayoutBox to handle. However, there was code on the layout
side that depended on IsFloating() getting adjusted for the style, as
well. This resulted in the DCHEK mentioned above getting hit.

To fix this, I updated all cases that used the style's IsFloating()
function and either updated it to also take the style's
IsFlexOrGridItem() function into account or updated the code to use
LayoutObject's IsFloating() function instead.

-----

Original change's description:
> Don't reset computed 'float' for flex and grid items
>
> Floated flex and grid items have their 'float' incorrectly computed to
> "none". This change fixes this by adjusting IsFloating() for flex and
> grid items inside LayoutBox::UpdateFromStyle() instead of inside
> StyleAdjuster::AdjustStyleForDisplay().
>
> Beyond this, legend elements are not allowed to be rendered legends
> if they are floating. However, if legends are flex items, we
> adjust IsFloating() to be false. This causes legends to be
> used as rendered legends when they are flex items, even if they
> were styled as a float.
>
> This is fixed by checking ComputedStyle::IsFloating() instead
> of LayoutObject::IsFloating() when finding the rendered legend in
> order to get its non-adjusted floating value.
>
> Bug: 875235,350505
> Change-Id: Ia1e7f7c244cc0c443cd58be42854866884b7f7e7
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2157889
> Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
> Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
> Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
> Commit-Queue: Alison Maher <almaher@microsoft.com>
> Cr-Commit-Position: refs/heads/master@{#763988}

Bug: 875235, 350505
Change-Id: I915e045db7fda94d99cfdae33911d8b5f6df26de